+++ src/interfaces/Interfaces.jl
@@ -2,34 +2,31 @@ module Interfaces
 
 import ..Frames
 
-"""
-Abstract type for Interfaces.
-"""
-abstract type AbstractCANInterface end
+include("InterfaceDef.jl")
 
 """
     send(interface::T<:AbstractCANInterface, frame::AbstractFrame)
 
-Abstract function for send message.
+function for send message.
 
-Common behavior of concrete implements:
+It behaves:
 * When send successed, return nothing.
 * When send failed, throws error.
 * Classic CAN interfaces can send ONLY `Frame`
 * CAN FD interfaces can send both `Frame` and `FDFrame`
 """
-function send()
-    error("abstract 'send' is not implemented.")
+function send(iface::Interface, frm::Frames.AnyFrame)
+    send(iface.driver, frm)
 end
 
 """
     recv(interface::T<:AbstractCANInterface; timeout_s::Real=0)
 
-Abstract function for receive message.
+function for receive message.
 
-Common behavior of concrete implements:
+It behaves:
 * Default non-blocking.
-    * For blocking receivement, set `timeout_s` in seconds. 
+    * For blocking receivement, set kwarg `timeout_s` in seconds. 
     * Set `timeout_s` < 0 for infinite bloking.
 * When receive successed, returns `Frame` or `FDFrame`.
 * When receive queue is empty, returns nothing.
@@ -37,44 +34,28 @@ Common behavior of concrete implements:
 * Classic CAN interfaces return only `Frame` object (except `slcan`).
 * CAN FD interfaces return either `Frame` or `FDFrame` object.
 """
-function recv()
-    error("abstract 'recv' is not implemented.")
+function recv(iface::Interface; timeout_s::Real=0)
+    recv(iface.driver; timeout_s)
 end
 
 
 """
     shutdown(interface::T<:AbstractCANInterface)
 
-Abstract function for shutdown interface.
+function for shutdown interface.
 Always returns nothing.
 """
-function shutdown()
-    error("abstract 'shutdown' is not implemented.")
+function shutdown(iface::Interface)
+    shutdown(iface.driver)
 end
 
 
-"""
-    AcceptanceFilter(code_id, mask)
-
-Struct for accept filter. 
-    
-If this struct is set to Interface,
-the id is accepted when 
-`<received_id> & mask == code_id & mask`
-or
-`(<received_id> xor code_id) & mask == 0 `.
-Those are equivalent.
-"""
-struct AcceptanceFilter
-    code_id::UInt32
-    mask::UInt32
-end
 
 
 include("vector/Vector.jl")
-include("kvaser/Kvaser.jl")
-include("socketcan/Socketcan.jl")
-include("slcan/Slcan.jl")
+# include("kvaser/Kvaser.jl")
+# include("socketcan/Socketcan.jl")
+# include("slcan/Slcan.jl")
 
 
 end # Interfaces