+++ docs/src/changelog.md
@@ -6,8 +6,8 @@ first beta release.
 
 * Support do-end statement.
 * Add automatic bittiming calculation from samplepoint.
-* tested on physical hardware (Vector, Kvaser).
-* bug fix: Notification object infinitely reproduce in Vector timeout.
+* Tested on physical hardware (Vector, Kvaser).
+* bug fix: Timeout notification object is infinitely reproduced in Vector.
 
 ### v0.0.5